requirements.txt
streamlit
pdfplumber
Flask
google-generativeai
json
